requirements.txt
grpcio
grpcio-reflection
-i https://download.pytorch.org/whl/cu118
torch
langchain
transformers==4.31.0
bitsandbytes-windows
accelerate
xformers
einops
scipy
sentencepiece
jsonformer
auto-gptq
confluent-kafka

python-dotenv~=1.0.0
pymongo~=4.4.1